requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
azure-search-documents==11.4.0b11
azure-identity==1.15.0
msal==1.24.1
msal-extensions==1.0.0
numpy==1.26.1 
openai==1.1.1
langchain==0.1.8
tiktoken>=0.5.2
python-dotenv>=1.0.0
PyPDF2==3.0.1 
requests>=2.31.0
redis>=5.0.1
pysolr>=3.9.0 